refactor(dh): move util function in ui library
