+++ libs/dh/metering-point/ui-status-badge/src/lib/connection-state-to-badge-type.spec.ts
