+++ libs/dh/metering-point/ui-status-badge/src/lib/dh-metering-point-status-badge.component.ts
@@ -25,7 +25,8 @@ import { TranslocoModule } from '@ngneat/transloco';
 
 import { ConnectionState } from '@energinet-datahub/dh/shared/data-access-api';
 import { WattBadgeModule, WattBadgeType } from '@energinet-datahub/watt';
-import { connectionStateToBadgeType } from '@energinet-datahub/dh/metering-point/domain';
+
+import { connectionStateToBadgeType } from './connection-state-to-badge-type';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
